requirements.txt
datasets 
transformers 
faiss-cpu